# openapi.json
{"openapi": "3.1.0", "info": {"title": "FastAPI", "version": "0.1.0"}, "paths": {"/v1/organizations/": {"post": {"tags": ["Organizations"], "summary": "Create Organization", "operationId": "create_organization_v1_organizations__post", "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/OrganizationCreate"}}}}, "responses": {"201": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Organization"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "get": {"tags": ["Organizations"], "summary": "List Organizations", "operationId": "list_organizations_v1_organizations__get", "parameters": [{"name": "limit", "in": "query", "required": false, "schema": {"type": "integer", "maximum": 100, "exclusiveMinimum": 0, "default": 100, "title": "Limit"}}, {"name": "offset", "in": "query", "required": false, "schema": {"type": "integer", "minimum": 0, "default": 0, "title": "Offset"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/PaginatedResponse_Organization_"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/v1/organizations/{organization_id}": {"get": {"tags": ["Organizations"], "summary": "Get Organization", "operationId": "get_organization_v1_organizations__organization_id__get", "parameters": [{"name": "organization_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Organization Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Organization"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "put": {"tags": ["Organizations"], "summary": "Update Organization", "operationId": "update_organization_v1_organizations__organization_id__put", "parameters": [{"name": "organization_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Organization Id"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/OrganizationUpdate"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Organization"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/v1/users/": {"post": {"tags": ["Users"], "summary": "Create User", "operationId": "create_user_v1_users__post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/UserCreate"}}}, "required": true}, "responses": {"201": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/User"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/v1/users/{user_id}": {"get": {"tags": ["Users"], "summary": "Get User", "operationId": "get_user_v1_users__user_id__get", "parameters": [{"name": "user_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid", "title": "User Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/User"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "put": {"tags": ["Users"], "summary": "Update User", "operationId": "update_user_v1_users__user_id__put", "parameters": [{"name": "user_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid", "title": "User Id"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/UserUpdate"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/User"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/v1/organizations/{organization_id}/users/": {"get": {"tags": ["Users"], "summary": "List Organization Users", "operationId": "list_organization_users_v1_organizations__organization_id__users__get", "parameters": [{"name": "organization_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Organization Id"}}, {"name": "limit", "in": "query", "required": false, "schema": {"type": "integer", "maximum": 100, "exclusiveMinimum": 0, "default": 100, "title": "Limit"}}, {"name": "offset", "in": "query", "required": false, "schema": {"type": "integer", "minimum": 0, "default": 0, "title": "Offset"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/PaginatedResponse_User_"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/v1/organizations/{organization_id}/users/{user_id}/make-admin": {"post": {"tags": ["Users"], "summary": "User Make Admin", "operationId": "user_make_admin_v1_organizations__organization_id__users__user_id__make_admin_post", "parameters": [{"name": "organization_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Organization Id"}}, {"name": "user_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid", "title": "User Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/User"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/v1/entitlements/": {"post": {"tags": ["Entitlements"], "summary": "Create Entitlement", "operationId": "create_entitlement_v1_entitlements__post", "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/EntitlementCreate"}}}}, "responses": {"201": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Entitlement"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "get": {"tags": ["Entitlements"], "summary": "List Entitlements", "operationId": "list_entitlements_v1_entitlements__get", "parameters": [{"name": "limit", "in": "query", "required": false, "schema": {"type": "integer", "maximum": 100, "exclusiveMinimum": 0, "default": 100, "title": "Limit"}}, {"name": "offset", "in": "query", "required": false, "schema": {"type": "integer", "minimum": 0, "default": 0, "title": "Offset"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/PaginatedResponse_Entitlement_"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/v1/entitlements/{entitlement_id}": {"get": {"tags": ["Entitlements"], "summary": "Get Entitlement", "operationId": "get_entitlement_v1_entitlements__entitlement_id__get", "parameters": [{"name": "entitlement_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Entitlement Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Entitlement"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "put": {"tags": ["Entitlements"], "summary": "Update Entitlement", "operationId": "update_entitlement_v1_entitlements__entitlement_id__put", "parameters": [{"name": "entitlement_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Entitlement Id"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/EntitlementUpdate"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Entitlement"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/v1/entitlements/{entitlement_id}/terminate": {"post": {"tags": ["Entitlements"], "summary": "Terminate Entitlement", "operationId": "terminate_entitlement_v1_entitlements__entitlement_id__terminate_post", "parameters": [{"name": "entitlement_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Entitlement Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Entitlement"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/v1/data-sources/": {"get": {"tags": ["Data Sources"], "summary": "List Data Sources", "operationId": "list_data_sources_v1_data_sources__get", "parameters": [{"name": "limit", "in": "query", "required": false, "schema": {"type": "integer", "maximum": 100, "exclusiveMinimum": 0, "default": 100, "title": "Limit"}}, {"name": "offset", "in": "query", "required": false, "schema": {"type": "integer", "minimum": 0, "default": 0, "title": "Offset"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/PaginatedResponse_DataSource_"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/v1/data-sources/{data_source_id}": {"get": {"tags": ["Data Sources"], "summary": "Get Data Source", "operationId": "get_data_source_v1_data_sources__data_source_id__get", "parameters": [{"name": "data_source_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Data Source Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/DataSource"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/v1/organizations/{organization_id}/data-sources/": {"get": {"tags": ["Data Sources"], "summary": "List Organization Data Sources", "operationId": "list_organization_data_sources_v1_organizations__organization_id__data_sources__get", "parameters": [{"name": "organization_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Organization Id"}}, {"name": "limit", "in": "query", "required": false, "schema": {"type": "integer", "maximum": 100, "exclusiveMinimum": 0, "default": 100, "title": "Limit"}}, {"name": "offset", "in": "query", "required": false, "schema": {"type": "integer", "minimum": 0, "default": 0, "title": "Offset"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/PaginatedResponse_DataSource_"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}}, "components": {"schemas": {"DataSource": {"properties": {"organisationId": {"type": "string", "format": "uuid", "title": "Organisationid", "description": "ID of the organization this data source belongs to", "examples": ["123e4567-e89b-12d3-a456-426614174000"]}, "type": {"$ref": "#/components/schemas/DataSourceType", "description": "Type of the data source", "examples": ["aws_root"]}, "resourcesChangedThisMonth": {"type": "integer", "title": "Resourceschangedthismonth", "description": "Number of resources that changed during the current month", "examples": [42]}, "expensesSoFarThisMonth": {"type": "string", "title": "Expensessofarthismonth", "description": "Current month's expenses up to now", "examples": ["1234.56"]}, "expensesForecastThisMonth": {"type": "string", "title": "Expensesforecastthismonth", "description": "Forecasted expenses for the current month", "examples": ["2500.00"]}, "iconUrl": {"type": "string", "maxLength": 2083, "minLength": 1, "format": "uri", "title": "Iconurl", "description": "URL to the icon representing this data source", "examples": ["https://example.com/icons/aws.png"]}}, "type": "object", "required": ["organisationId", "type", "resourcesChangedThisMonth", "expensesSoFarThisMonth", "expensesForecastThisMonth", "iconUrl"], "title": "DataSource"}, "DataSourceType": {"type": "string", "enum": ["aws_root", "aws_linked", "azure_tenant", "azure_subscription", "gcp"], "title": "DataSourceType"}, "Entitlement": {"properties": {"sponsorName": {"type": "string", "title": "Sponsorname", "description": "Name of the sponsor for this entitlement", "examples": ["AWS"]}, "sponsorExternalId": {"type": "string", "title": "Sponsorexternalid", "description": "Vendor account number from MPT", "examples": ["ACC-1234-5678"]}, "sponsorContainerId": {"type": "string", "title": "Sponsorcontainerid", "description": "Azure Sub ID, AWS Account number, GCP Project ID etc.", "examples": ["d4500e78-ac78-4445-89e8-5b8a4d482035"]}, "entitlementId": {"type": "string", "format": "uuid", "title": "Entitlementid", "description": "Unique identifier for the entitlement", "examples": ["123e4567-e89b-12d3-a456-426614174000"]}, "activatedAt": {"anyOf": [{"type": "string", "format": "date-time"}, {"type": "null"}], "title": "Activatedat", "description": "Timestamp when the entitlement was activated"}, "activatedBy": {"anyOf": [{"type": "string", "format": "uuid"}, {"type": "null"}], "title": "Activatedby", "description": "User ID who activated the entitlement"}, "terminatedAt": {"anyOf": [{"type": "string", "format": "date-time"}, {"type": "null"}], "title": "Terminatedat", "description": "Timestamp when the entitlement was terminated", "examples": ["2023-12-31T12:00:00Z"]}, "terminatedBy": {"anyOf": [{"type": "string", "format": "uuid"}, {"type": "null"}], "title": "Terminatedby", "description": "User ID who terminated the entitlement"}}, "type": "object", "required": ["sponsorName", "sponsorExternalId", "sponsorContainerId", "entitlementId", "activatedAt", "activatedBy", "terminatedAt", "terminatedBy"], "title": "Entitlement"}, "EntitlementCreate": {"properties": {"sponsorName": {"type": "string", "title": "Sponsorname", "description": "Name of the sponsor for this entitlement", "examples": ["AWS"]}, "sponsorExternalId": {"type": "string", "title": "Sponsorexternalid", "description": "Vendor account number from MPT", "examples": ["ACC-1234-5678"]}, "sponsorContainerId": {"type": "string", "title": "Sponsorcontainerid", "description": "Azure Sub ID, AWS Account number, GCP Project ID etc.", "examples": ["d4500e78-ac78-4445-89e8-5b8a4d482035"]}}, "type": "object", "required": ["sponsorName", "sponsorExternalId", "sponsorContainerId"], "title": "EntitlementCreate"}, "EntitlementUpdate": {"properties": {"sponsorName": {"type": "string", "title": "Sponsorname", "description": "Name of the sponsor for this entitlement", "examples": ["AWS"]}, "sponsorExternalId": {"type": "string", "title": "Sponsorexternalid", "description": "Vendor account number from MPT", "examples": ["ACC-1234-5678"]}, "sponsorContainerId": {"type": "string", "title": "Sponsorcontainerid", "description": "Azure Sub ID, AWS Account number, GCP Project ID etc.", "examples": ["d4500e78-ac78-4445-89e8-5b8a4d482035"]}}, "type": "object", "required": ["sponsorName", "sponsorExternalId", "sponsorContainerId"], "title": "EntitlementUpdate"}, "HTTPValidationError": {"properties": {"detail": {"items": {"$ref": "#/components/schemas/ValidationError"}, "type": "array", "title": "Detail"}}, "type": "object", "title": "HTTPValidationError"}, "Organization": {"properties": {"name": {"type": "string", "title": "Name", "description": "The name of the organization", "examples": ["Apple Inc."]}, "currency": {"type": "string", "enum": ["AED", "AFN", "ALL", "AMD", "ANG", "AOA", "ARS", "AUD", "AWG", "AZN", "BAM", "BBD", "BDT", "BGN", "BHD", "BIF", "BMD", "BND", "BOB", "BOV", "BRL", "BSD", "BTN", "BWP", "BYN", "BZD", "CAD", "CDF", "CHE", "CHF", "CHW", "CLF", "CLP", "CNY", "COP", "COU", "CRC", "CUC", "CUP", "CVE", "CZK", "DJF", "DKK", "DOP", "DZD", "EGP", "ERN", "ETB", "EUR", "FJD", "FKP", "GBP", "GEL", "GHS", "GIP", "GMD", "GNF", "GTQ", "GYD", "HKD", "HNL", "HRK", "HTG", "HUF", "IDR", "ILS", "INR", "IQD", "IRR", "ISK", "JMD", "JOD", "JPY", "KES", "KGS", "KHR", "KMF", "KPW", "KRW", "KWD", "KYD", "KZT", "LAK", "LBP", "LKR", "LRD", "LSL", "LYD", "MAD", "MDL", "MGA", "MKD", "MMK", "MNT", "MOP", "MRU", "MUR", "MVR", "MWK", "MXN", "MXV", "MYR", "MZN", "NAD", "NGN", "NIO", "NOK", "NPR", "NZD", "OMR", "PAB", "PEN", "PGK", "PHP", "PKR", "PLN", "PYG", "QAR", "RON", "RSD", "RUB", "RWF", "SAR", "SBD", "SCR", "SDG", "SEK", "SGD", "SHP", "SLE", "SLL", "SOS", "SRD", "SSP", "STN", "SVC", "SYP", "SZL", "THB", "TJS", "TMT", "TND", "TOP", "TRY", "TTD", "TWD", "TZS", "UAH", "UGX", "USD", "USN", "UYI", "UYU", "UYW", "UZS", "VED", "VES", "VND", "VUV", "WST", "XAF", "XCD", "XOF", "XPF", "XSU", "XUA", "YER", "ZAR", "ZMW", "ZWL"], "maxLength": 3, "minLength": 3, "title": "Currency", "description": "The primary currency used by the organization for financial operations", "examples": ["USD"]}, "id": {"type": "string", "format": "uuid", "title": "Id"}, "limit": {"type": "string", "title": "Limit", "description": "Maximum spending limit set for the organization", "examples": ["50000.00"]}, "expensesLastMonth": {"type": "string", "title": "Expenseslastmonth", "description": "Total expenses from the previous month", "examples": ["42350.75"]}, "expensesThisMonth": {"type": "string", "title": "Expensesthismonth", "description": "Current month's accumulated expenses", "examples": ["23150.25"]}, "expensesMonthForecast": {"type": "string", "title": "Expensesmonthforecast", "description": "Predicted total expenses for the current month", "examples": ["45000.00"]}, "possibleMonthlySavings": {"type": "string", "title": "Possiblemonthlysavings", "description": "Estimated amount that could be saved based on current spending patterns", "examples": ["5000.00"]}}, "type": "object", "required": ["name", "currency", "id", "limit", "expensesLastMonth", "expensesThisMonth", "expensesMonthForecast", "possibleMonthlySavings"], "title": "Organization"}, "OrganizationCreate": {"properties": {"name": {"type": "string", "title": "Name", "description": "The name of the organization", "examples": ["Apple Inc."]}, "currency": {"type": "string", "enum": ["AED", "AFN", "ALL", "AMD", "ANG", "AOA", "ARS", "AUD", "AWG", "AZN", "BAM", "BBD", "BDT", "BGN", "BHD", "BIF", "BMD", "BND", "BOB", "BOV", "BRL", "BSD", "BTN", "BWP", "BYN", "BZD", "CAD", "CDF", "CHE", "CHF", "CHW", "CLF", "CLP", "CNY", "COP", "COU", "CRC", "CUC", "CUP", "CVE", "CZK", "DJF", "DKK", "DOP", "DZD", "EGP", "ERN", "ETB", "EUR", "FJD", "FKP", "GBP", "GEL", "GHS", "GIP", "GMD", "GNF", "GTQ", "GYD", "HKD", "HNL", "HRK", "HTG", "HUF", "IDR", "ILS", "INR", "IQD", "IRR", "ISK", "JMD", "JOD", "JPY", "KES", "KGS", "KHR", "KMF", "KPW", "KRW", "KWD", "KYD", "KZT", "LAK", "LBP", "LKR", "LRD", "LSL", "LYD", "MAD", "MDL", "MGA", "MKD", "MMK", "MNT", "MOP", "MRU", "MUR", "MVR", "MWK", "MXN", "MXV", "MYR", "MZN", "NAD", "NGN", "NIO", "NOK", "NPR", "NZD", "OMR", "PAB", "PEN", "PGK", "PHP", "PKR", "PLN", "PYG", "QAR", "RON", "RSD", "RUB", "RWF", "SAR", "SBD", "SCR", "SDG", "SEK", "SGD", "SHP", "SLE", "SLL", "SOS", "SRD", "SSP", "STN", "SVC", "SYP", "SZL", "THB", "TJS", "TMT", "TND", "TOP", "TRY", "TTD", "TWD", "TZS", "UAH", "UGX", "USD", "USN", "UYI", "UYU", "UYW", "UZS", "VED", "VES", "VND", "VUV", "WST", "XAF", "XCD", "XOF", "XPF", "XSU", "XUA", "YER", "ZAR", "ZMW", "ZWL"], "maxLength": 3, "minLength": 3, "title": "Currency", "description": "The primary currency used by the organization for financial operations", "examples": ["USD"]}}, "type": "object", "required": ["name", "currency"], "title": "OrganizationCreate"}, "OrganizationUpdate": {"properties": {"name": {"type": "string", "title": "Name", "description": "The name of the organization", "examples": ["Apple Inc."]}, "currency": {"type": "string", "enum": ["AED", "AFN", "ALL", "AMD", "ANG", "AOA", "ARS", "AUD", "AWG", "AZN", "BAM", "BBD", "BDT", "BGN", "BHD", "BIF", "BMD", "BND", "BOB", "BOV", "BRL", "BSD", "BTN", "BWP", "BYN", "BZD", "CAD", "CDF", "CHE", "CHF", "CHW", "CLF", "CLP", "CNY", "COP", "COU", "CRC", "CUC", "CUP", "CVE", "CZK", "DJF", "DKK", "DOP", "DZD", "EGP", "ERN", "ETB", "EUR", "FJD", "FKP", "GBP", "GEL", "GHS", "GIP", "GMD", "GNF", "GTQ", "GYD", "HKD", "HNL", "HRK", "HTG", "HUF", "IDR", "ILS", "INR", "IQD", "IRR", "ISK", "JMD", "JOD", "JPY", "KES", "KGS", "KHR", "KMF", "KPW", "KRW", "KWD", "KYD", "KZT", "LAK", "LBP", "LKR", "LRD", "LSL", "LYD", "MAD", "MDL", "MGA", "MKD", "MMK", "MNT", "MOP", "MRU", "MUR", "MVR", "MWK", "MXN", "MXV", "MYR", "MZN", "NAD", "NGN", "NIO", "NOK", "NPR", "NZD", "OMR", "PAB", "PEN", "PGK", "PHP", "PKR", "PLN", "PYG", "QAR", "RON", "RSD", "RUB", "RWF", "SAR", "SBD", "SCR", "SDG", "SEK", "SGD", "SHP", "SLE", "SLL", "SOS", "SRD", "SSP", "STN", "SVC", "SYP", "SZL", "THB", "TJS", "TMT", "TND", "TOP", "TRY", "TTD", "TWD", "TZS", "UAH", "UGX", "USD", "USN", "UYI", "UYU", "UYW", "UZS", "VED", "VES", "VND", "VUV", "WST", "XAF", "XCD", "XOF", "XPF", "XSU", "XUA", "YER", "ZAR", "ZMW", "ZWL"], "maxLength": 3, "minLength": 3, "title": "Currency", "description": "The primary currency used by the organization for financial operations", "examples": ["USD"]}}, "type": "object", "required": ["name", "currency"], "title": "OrganizationUpdate"}, "PaginatedResponse_DataSource_": {"properties": {"pagination": {"$ref": "#/components/schemas/PaginationData"}, "items": {"items": {"$ref": "#/components/schemas/DataSource"}, "type": "array", "title": "Items", "description": "List of items for the current page"}}, "type": "object", "required": ["pagination", "items"], "title": "PaginatedResponse[DataSource]"}, "PaginatedResponse_Entitlement_": {"properties": {"pagination": {"$ref": "#/components/schemas/PaginationData"}, "items": {"items": {"$ref": "#/components/schemas/Entitlement"}, "type": "array", "title": "Items", "description": "List of items for the current page"}}, "type": "object", "required": ["pagination", "items"], "title": "PaginatedResponse[Entitlement]"}, "PaginatedResponse_Organization_": {"properties": {"pagination": {"$ref": "#/components/schemas/PaginationData"}, "items": {"items": {"$ref": "#/components/schemas/Organization"}, "type": "array", "title": "Items", "description": "List of items for the current page"}}, "type": "object", "required": ["pagination", "items"], "title": "PaginatedResponse[Organization]"}, "PaginatedResponse_User_": {"properties": {"pagination": {"$ref": "#/components/schemas/PaginationData"}, "items": {"items": {"$ref": "#/components/schemas/User"}, "type": "array", "title": "Items", "description": "List of items for the current page"}}, "type": "object", "required": ["pagination", "items"], "title": "PaginatedResponse[User]"}, "PaginationData": {"properties": {"limit": {"type": "integer", "maximum": 100.0, "exclusiveMinimum": 0.0, "title": "Limit", "description": "Number of records to return", "default": 100}, "offset": {"type": "integer", "minimum": 0.0, "title": "Offset", "description": "Number of records to skip", "default": 0}, "total": {"type": "integer", "minimum": 0.0, "title": "Total", "description": "Total number of records available"}}, "type": "object", "required": ["total"], "title": "PaginationData"}, "User": {"properties": {"email": {"type": "string", "format": "email", "title": "Email"}, "id": {"type": "string", "format": "uuid", "title": "Id"}, "organisationId": {"type": "string", "format": "uuid", "title": "Organisationid"}, "role": {"$ref": "#/components/schemas/UserRole"}, "createdAt": {"type": "string", "format": "date-time", "title": "Createdat"}, "lastLoginAt": {"type": "string", "format": "date-time", "title": "Lastloginat"}}, "type": "object", "required": ["email", "id", "organisationId", "role", "createdAt", "lastLoginAt"], "title": "User"}, "UserCreate": {"properties": {"email": {"type": "string", "format": "email", "title": "Email"}, "organisationId": {"type": "string", "format": "uuid", "title": "Organisationid"}, "role": {"$ref": "#/components/schemas/UserRole"}}, "type": "object", "required": ["email", "organisationId", "role"], "title": "UserCreate"}, "UserRole": {"type": "string", "enum": ["organisation_manager", "manager", "engineer"], "title": "UserRole"}, "UserUpdate": {"properties": {"email": {"type": "string", "format": "email", "title": "Email"}}, "type": "object", "required": ["email"], "title": "UserUpdate"}, "ValidationError": {"properties": {"loc": {"items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}, "type": "array", "title": "Location"}, "msg": {"type": "string", "title": "Message"}, "type": {"type": "string", "title": "Error Type"}}, "type": "object", "required": ["loc", "msg", "type"], "title": "ValidationError"}}}}